+++ R/plot.R
@@ -50,7 +50,11 @@ plot.marginaleffects <- function(x,
              ggplot2::labs(x = xlab, y = "")
     }
 
-    # theme and return
-    p <- p + ggplot2::theme_minimal()
-    return(p) 
+    # set a new theme only if the default is theme_grey. this prevents user's
+    # theme_set() from being overwritten
+    if (identical(ggplot2::theme_get(), ggplot2::theme_grey())) {
+        p <- p + ggplot2::theme_minimal()
+    }
+
+    return(p)
 }