+++ NEWS.md
@@ -1,4 +1,4 @@
-# marginaleffects 0.3.5
+# marginaleffects 0.4.0
 
 New function:
 
@@ -7,7 +7,7 @@ New function:
 Misc:
 
 * Speed optimizations
-* `predictions()` and `plot_cap()` includes confidence intervals for linear models
+* `predictions()` and `plot_cap()` include confidence intervals for linear models
 * More robust handling of in-formula functions: factor(), strata(), mo()
 * Do not overwrite user's `ggplot2::theme_set()` call